Move graph searching to GraphtUtils
